* alpha.c (print_operand_address): Fix format specifier warnings. 	* alpha/elf.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise. 	* alpha/vms.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise. 	* arm/aof.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL): Likewise. 	* arm/pe.h (ASM_OUTPUT_COMMON): Likewise. 	* avr.h (ASM_OUTPUT_COMMON, ASM_GENERATE_INTERNAL_LABEL, 	ASM_OUTPUT_SKIP): Likewise. 	* c4x.h (ASM_GENERATE_INTERNAL_LABEL): Likewise. 	* dsp16xx.h (ASM_GENERATE_INTERNAL_LABEL): Likewise. 	* h8300.h (ASM_GENERATE_INTERNAL_LABEL, ASM_OUTPUT_COMMON): 	Likewise. 	* i370.h (ASM_GENERATE_INTERNAL_LABEL): Likewise. 	* i386/cygming.h (ASM_OUTPUT_COMMON): Likewise. 	* i386/darwin.h (ASM_OUTPUT_COMMON): Likewise. 	* i960.h (ASM_GENERATE_INTERNAL_LABEL): Likewise. 	* m68k/hp320.h (PRINT_OPERAND_ADDRESS): Likewise. 	* mcore.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise. 	* pdp11.h (ASM_GENERATE_INTERNAL_LABEL): Likewise. 	* ptx4.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise. 	* sparc/freebsd.h (ASM_GENERATE_INTERNAL_LABEL): Likewise. 	* svr3.h (ASM_OUTPUT_COMMON): Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@67453 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/config/arm/aof.h
@@ -207,14 +207,14 @@ do					\
    fprintf ((STREAM), "\tAREA "),				\
    assemble_name ((STREAM), (NAME)),				\
    fprintf ((STREAM), ", DATA, COMMON\n\t%% %d\t%s size=%d\n",	\
-	    (ROUNDED), ASM_COMMENT_START, (int)SIZE))
+	    (int)(ROUNDED), ASM_COMMENT_START, (int)(SIZE)))
 
 #define ASM_OUTPUT_LOCAL(STREAM,NAME,SIZE,ROUNDED)	\
    (zero_init_section (),				\
     assemble_name ((STREAM), (NAME)),			\
     fprintf ((STREAM), "\n"),				\
     fprintf ((STREAM), "\t%% %d\t%s size=%d\n",		\
-	     (ROUNDED), ASM_COMMENT_START, (int)SIZE))
+	     (int)(ROUNDED), ASM_COMMENT_START, (int)(SIZE)))
 
 /* Output and Generation of Labels */
